refactor: implement terminal size context provider to fix max listeners warning
