+++ src/ui/App.tsx
@@ -13,6 +13,7 @@ import { Markdown } from './Markdown';
 import { Messages } from './Messages';
 import { QueueDisplay } from './QueueDisplay';
 import { useAppStore } from './store';
+import { TerminalSizeProvider } from './TerminalSizeContext';
 import { TranscriptModeIndicator } from './TranscriptModeIndicator';
 import { useTerminalRefresh } from './useTerminalRefresh';
 
@@ -127,31 +128,33 @@ export function App() {
     })();
   }, [forkModalVisible, bridge, cwd, sessionId]);
   return (
-    <Box
-      flexDirection="column"
-      key={`${forceRerender}-${forkParentUuid}-${forkCounter}-${transcriptMode}`}
-    >
-      <Messages />
-      <BackgroundPrompt />
-      <PlanResult />
-      <ActivityIndicator />
-      <QueueDisplay />
-      {transcriptMode ? <TranscriptModeIndicator /> : <ChatInput />}
-      <SlashCommandJSX />
-      <ApprovalModal />
-      {forkModalVisible && (
-        <ForkModal
-          messages={forkMessages as any}
-          onSelect={(uuid) => {
-            fork(uuid);
-          }}
-          onClose={() => {
-            hideForkModal();
-          }}
-        />
-      )}
-      <ExitHint />
-      <Debug />
-    </Box>
+    <TerminalSizeProvider>
+      <Box
+        flexDirection="column"
+        key={`${forceRerender}-${forkParentUuid}-${forkCounter}-${transcriptMode}`}
+      >
+        <Messages />
+        <BackgroundPrompt />
+        <PlanResult />
+        <ActivityIndicator />
+        <QueueDisplay />
+        {transcriptMode ? <TranscriptModeIndicator /> : <ChatInput />}
+        <SlashCommandJSX />
+        <ApprovalModal />
+        {forkModalVisible && (
+          <ForkModal
+            messages={forkMessages as any}
+            onSelect={(uuid) => {
+              fork(uuid);
+            }}
+            onClose={() => {
+              hideForkModal();
+            }}
+          />
+        )}
+        <ExitHint />
+        <Debug />
+      </Box>
+    </TerminalSizeProvider>
   );
 }
